Introcude uniffi-bindgen-cs as library
This is required for 0.23 update, since the standard method for using
uniffi has changed. Now, instead of providing a binary, we should
provide a library with binary's main function, and the consumer should
wrap the library into a binary.

To allow multiple libraries (uniffi-bindgen-cs and
uniffi-bindgen-cs-fixtures), split package into 'bindgen' and
'fixtures' packages.

Signed-off-by: Kristupas Antanavicius <kristupas.antanavicius@nordsec.com>