Lowercase the numeric types
Lowercasing numeric types introduces a problem. In C# custom types are
implemented with `using` directive, and the `using` directive expects
and identifier on the right side of `=`. Lowercase numeric types are
not identifiers, but rather reserved keywords. So its not possible to
define a type alias using a lowercase numeric type as the underlying
type. To use numeric types as the underlying type, the uppercase
system type counterparts must be used

To generate correct code for custom types, create new filter
`type_name_custom`, that formats these types in their respective system
type counterparts.

Signed-off-by: Kristupas Antanavicius <kristupas.antanavicius@nordsec.com>