Fix error handling for command `dotnet csharpier`
In Rust, command exit status has to be handled manually. Not handling
exit status would cause the generator to generate invalid code when
`dotnet` is installed, but `csharpier` is not installed.

Switch from using `dotnet csharpier` to using `dotnet-csharpier`. The
former produces really verbose output by `dotnet` when `csharpier` is
not installed or dotnet tools directory is not in PATH, while the latter
returns a neat `command not found` error if CSharpier is not yet
installed.

Signed-off-by: Kristupas Antanavicius <kristupas.antanavicius@nordsec.com>