Update objects ABI to match 0.28.0
SafeHandle is used to manage ownership of the Rust object. When
SafeHandle goes out of scope, the underlying Rust reference is
decremented. Method calls and lowering increment ("clone") object
reference, and Rust is responsible for decrementing the reference. This
means, it's not necessary to use SafeHandle directly for method calls
and lowering, since Rust will take care of decrementing the reference.

However, it's still necessary to use SafeHandle for "cloning" the
reference, since CG may choose to collect the object itself and run the
finalizer, in turn decrementing Rust reference and causing
use-after-free during "clone" native call. (See SafeHandle
documentation).

Signed-off-by: Daniel Fetti <daniel.fetti@nordsec.com>