Execute tests in sequence (#102)
By default, the tests are executed in parallel. While this doesn't seem
to be causing any type of issues, it's quite counter intuitive for the
type of tests implemented in this project. For example, some of the
tests invoke GC and expect side effects from GC.

This commit also adds diagnostic messages to help to debug hanging
tests. I implemented this after encountering a hanging test in one of
Github PRs.

Signed-off-by: Kristupas Antanavičius <kristupas.antanavicius@nordsec.com>